#46: Renamed menu visitors to toolbar visitors
